+++ src/gatelin-express.ts
@@ -2,4 +2,14 @@
 export {
   getConsumer,
 } from './consumer';
+
+export {
+  getAcl,
+  stripUnallowedFields,
+} from './acl';
+
+export type {
+  Acl,
+  AclCondition,
+} from './acl';